Make easier to use as library
